Fix/Improve certain holidays rarely returning any songs
The offset for the Spotify search often resulted in holiday searches not
yielding any songs. The offset for holiday searches is not set to 0. This
might still be an issue for some holidays, where the search keyword simply
doesn't return any songs. To fully fix this a holiday to search word module
might be the answer, TODO.